Move hasExternalLinks to Entry file.
